tests: Move generic helpers to utils.py
